+++ src/util/Integrator.h
@@ -73,7 +73,7 @@ class Integrator{
        int blackhol;
        float inittol;
        int maxit;
-       long maxitmon;
+       int maxitmon;
        int retry;
 
        int syint;
